requirements.txt
thirdai[neural_db]
tika
openai
requests
nltk
pandas
numpy
python-dotenv
streamlit